Moving the `addColorBar` function to shared plotting module.
